+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * (core) removing need to call `Ionic.io();` manually
 * (deploy) update() can now defer reloading the app
+* (deploy) adding `getVersions` and `deleteVersion` methods to manage deploy versions.
 
 
 ## 0.1.1